Catalogue Class: BookId Validation was moved to be perfomed in Catalogue Class instead of Library Class. There was a bug where adding a book from the Catalouge class would allow multiple books to exist in the Catalogue collection with an identical ID number. Therefore when adding a book using the Library class method will now call the method in the Catalouge class and thus performing the validation automatically. The boolean isValidBookId(int id) method was moved to Catalouge class from Library Class. addBook will not throw an exception if a book with the same id already exists. The method will simply not add the book to the collection. getAllBooks will return the collection list as is.
Library Class: refer to Catalouge class changes

CatalougeTest Class: changed book object to make sure that the books are kept in the static list and for test suites that follow. This is verified in LibraryTest.java

LibraryTest Class: modified assertions to reflect the size of collection given that the book objects are carried forwared form the CatalougeTest Class. Gave naming conventions to test methods to ensure running order as desired by developers.